src/COMP0073_SmartVision_Project/requirements.txt
##### Requirements for the Proof of Concept SmartVision System #####
pillow
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-customvision
azure-cognitiveservices-vision-face
azure-storage-blob
mysql-connector-python
numpy
matplotlib
Flask
schedule
